More javascript refactoring:
- expost jquery as `$` using browserify-shim
- prepend all analyse-form-related modules with `analyse-`
- move all modules to `src/` directory for consistency
- use `domready` npm package instead of jquery `document.ready` idiom
- delete unused scatter-chart module
- only include javascript for specific pages on those pages
