+++ openprescribing/media/js/src/tariff-charts.js
@@ -1,9 +1,9 @@
-global.jQuery = require('jquery');
-global.$ = global.jQuery;
+var $ = require('jquery');
 require('bootstrap');
 require('select2');
 require('Highcharts');
 require('Highcharts-export');
+var domready = require("domready");
 var _ = require('underscore');
 
 var chartOptions = require('./highcharts-options');
@@ -263,6 +263,6 @@ var tariffChart = {
   },
 };
 
-$(document).ready(function() {
+domready(function() {
   tariffChart.setUp();
 });
