+++ openprescribing/templates/analyse.html
@@ -224,5 +224,6 @@
 <!--[if !IE 8]><!-->
   <script src="/static/js/clipboard.min.js?q=123456"></script>
 <!--<![endif]-->
-{% conditional_js 'openprescribing' %}
+{% conditional_js 'analyse-form' %}
+
 {% endblock %}
