+++ openprescribing/templates/bnf_section.html
@@ -53,5 +53,5 @@
 var filename = config.apiHost + "{% url 'total_spending' %}?format=json&code={{ page_id }}";
 var pageType = 'bnf-section';
 </script>
-{% conditional_js 'bar-chart' %}
+{% conditional_js 'bar-charts' %}
 {% endblock %}
