+++ openprescribing/templates/measure_for_all_ccgs.html
@@ -69,7 +69,9 @@ text
 {% endverbatim %}
 
 {% include '_get_in_touch.html' %}
+{% endblock %}
 
+{% block extra_js %}
 <script>
 var measureData = {
     'orgType': 'CCG',
