+++ openprescribing/templates/measure_for_one_ccg.html
@@ -64,6 +64,10 @@ numerator was composed of the following presentations:</p>
 
 {% include '_get_in_touch.html' %}
 
+{% endblock %}
+
+{% block extra_js %}
+
 <script>
 var measureData = {
     'orgId': '{{ ccg.code }}',
@@ -74,9 +78,6 @@ var measureData = {
 };
 </script>
 
-{% endblock %}
-
-{% block extra_js %}
 {% conditional_js 'measures' %}
 {% include '_measure_breakdown_table.html' %}
 {% endblock %}
