+++ openprescribing/templates/measure_for_practices_in_ccg.html
@@ -70,6 +70,10 @@ text
 
 {% include '_get_in_touch.html' %}
 
+{% endblock %}
+
+
+{% block extra_js %}
 <script>
 var measureData = {
     'orgId': '{{ ccg.code }}',
