+++ openprescribing/templates/measures_for_one_ccg.html
@@ -49,13 +49,13 @@
 </div>
 
 <hr/>
-
 {% include '_measures_panel.html' with entity_type='CCG' %}
 
 <hr/>
-
 {% include '_get_in_touch.html' %}
+{% endblock %}
 
+{% block extra_js %}
 <script>
  var measureData = {
     'tags': '{{ request.GET.tags|default:"core" }}',
@@ -65,5 +65,4 @@
 };
 </script>
 {% conditional_js 'measures' %}
-
 {% endblock %}
